kis_trader/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
apscheduler==3.10.4
requests==2.32.3
pandas==2.2.3
ta==0.11.0
scikit-learn==1.6.1
joblib==1.4.2
cryptography==44.0.2
python-dotenv==1.0.1
